requirements.txt
absl-py==0.6.1
astor==0.7.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==3.1.0
boto3==1.12.1
botocore==1.15.1
cachetools==3.1.1
cffi==1.12.2
cmake==3.16.3
colorama==0.4.1
configparser==3.7.4
cycler==0.10.0
Cython==0.29.2
decorator==4.4.1
defusedxml==0.6.0
docopt==0.6.2
docutils==0.15.2
easydict==1.9
entrypoints==0.3
enum34==1.1.6
funcsigs==1.0.2
gast==0.2.0
gitdb2==2.0.6
GitPython==3.0.5
google==2.0.2
google-auth==1.4.2
greenlet==0.4.15
grpcio==1.16.1
h5py==2.10.0
imageio==2.6.1
imageio-ffmpeg==0.3.0
importlib-metadata==0.23
intel-openmp==2018.0.3
ipykernel==5.1.3
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
jmespath==0.9.4
jsonpickle==0.9.6
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
jupyter-http-over-ws==0.0.6
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.1.0
lap==0.4.0
lapsolver==1.0.2
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mock==2.0.0
more-itertools==7.2.0
motmetrics==1.1.3
msgpack==0.6.0
munch==2.5.0
munkres==1.1.2
nbconvert==5.6.1
nbformat==4.4.0
neovim==0.3.1
networkx==2.4
notebook==6.0.2
numpy==1.17.4
opencv-python==4.2.0.32
packaging==19.2
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pbr==5.4.4
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prometheus-client==0.7.1
prompt-toolkit==1.0.18
protobuf==3.11.3
psycopg2==2.8.4
ptyprocess==0.6.0
py-cpuinfo==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
Pygments==2.4.2
pynvim==0.3.1
pynvml==8.0.4
pyparsing==2.4.5
pyrsistent==0.15.5
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
pyzmq==18.1.1
qtconsole==4.5.5
rsa==4.0
s3transfer==0.3.3
sacred==0.8.0
scikit-image==0.16.2
scikit-learn==0.20.0
scipy==1.3.2
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.12.0
smmap2==2.0.5
soupsieve==1.9.5
SQLAlchemy==1.3.13
tensorboard==1.12.0
tensorboardX==2.0
tensorflow-gpu==1.1.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
torch==1.3.1
torchvision==0.4.2
tornado==6.0.3
tqdm==4.38.0
traitlets==4.3.3
trollius==2.1.post2
urllib3==1.25.8
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.3
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
